final parser production stabilization
Harden parser production behavior with forced proxy networking, queue and worker safeguards, robust health/preflight checks, progress speed migration fallback, and updated production setup documentation.

+++ app/Models/ParserProgress.php
@@ -21,7 +21,7 @@ class ParserProgress extends Model
         'total_items' => 'integer',
         'processed_items' => 'integer',
         'failed_items' => 'integer',
-        'speed_per_min' => 'float',
+        'speed_per_min' => 'integer',
     ];
 
     public function job(): BelongsTo
